refine aptos hasher types. refactor code structure to support many chains
