+++ src/lib/execution.ts
@@ -10,7 +10,7 @@ import {
 import type { TypedContract } from 'ethers-abitype'
 
 import Router from '../abi/Router.js'
-import { Tree, getLeafHasher, proofFlagsToBits } from './hasher/index.js'
+import { Tree, getLeafHasher, proofFlagsToBits } from './hasher'
 import {
   type CCIPContract,
   type CCIPExecution,
@@ -42,20 +42,20 @@ import {
  * @param merkleRoot - Optional merkleRoot of the CommitReport, for validation
  * @returns ManualExec report arguments
  **/
-export function calculateManualExecProof(
-  messagesInBatch: readonly CCIPMessage[],
-  lane: Lane,
+export function calculateManualExecProof<V extends CCIPVersion>(
+  messagesInBatch: readonly CCIPMessage<V>[],
+  lane: Lane<V>,
   messageIds: string[],
   merkleRoot?: string,
 ): {
-  messages: CCIPMessage[]
+  messages: CCIPMessage<V>[]
   proofs: string[]
   proofFlagBits: bigint
 } {
   const leaves: string[] = []
   const hasher = getLeafHasher(lane)
   const prove: number[] = []
-  const messages: CCIPMessage[] = []
+  const messages: CCIPMessage<V>[] = []
   const seen = new Set<string>()
 
   messagesInBatch.forEach((message, index) => {
@@ -71,10 +71,6 @@ export function calculateManualExecProof(
       )
     }
     // Hash leaf node
-    // TODO: msg here could be of any type (1.5 or 1.6), but we are using a specific version hasher. This will throw at runtime if encounters different message versions
-    // hasher type should detect the msg version
-    // If we want to be able to hash any message type, we should have a hasher per version
-    // If we don't want to hash any message type, `messagesInBatch` should only accept one type of message
     leaves.push(hasher(msg))
     const msgId = msg.header.messageId
     seen.add(msgId)
