+++ src/lib/hasher/index.ts
@@ -1,2 +1,3 @@
-export { type LeafHasher, ZERO_HASH, getLeafHasher, hashInternal } from './hasher.js'
+export { type LeafHasher, ZERO_HASH, hashInternal } from './common.js'
 export { MAX_NUMBER_TREE_LEAVES, Tree, proofFlagsToBits, verifyComputeRoot } from './merklemulti.js'
+export { getLeafHasher } from './hasher.js'
