+++ src/lib/selectors.ts
@@ -1,4 +1,6 @@
-const selectors: Record<number, { readonly selector: bigint; readonly name?: string }> = {
+type Selectors = Record<number, { readonly selector: bigint; readonly name?: string }>
+
+const evmSelectors: Selectors = {
   // generate:
   // fetch('https://github.com/smartcontractkit/chain-selectors/raw/main/selectors.yml')
   //   .then((res) => res.text())
@@ -382,4 +384,17 @@ const selectors: Record<number, { readonly selector: bigint; readonly name?: str
   },
   // end:generate
 }
+
+export const aptosSelectors: Selectors = {
+  '1': { name: 'aptos-mainnet', selector: 4741433654826277614n },
+  '2': { name: 'aptos-testnet', selector: 743186221051783445n },
+  '4': { name: 'aptos-localnet', selector: 4457093679053095497n },
+}
+
+const selectors: Selectors = { ...evmSelectors, ...aptosSelectors }
+
+export const isAptosChain = (selector: bigint): boolean => {
+  return Object.values(aptosSelectors).some((aptosSelector) => aptosSelector.selector === selector)
+}
+
 export default selectors
